radeon_bios.c: Allocate the ROM from the grand-father device
Using BUS_ALLOC_RESOURCE on the grand-father device allows to map the
device's Expansion ROM to be mapped. The returned start physical address
is then written to the ROM's BAR. After that, the ROM is available at
the virtual address.

If this fails, we still try with the shadow copy.

Suggested by:	kib@