radeon: In radeon_fence_wait_*seq(), don't return if r == EWOULDBLOCK
In these two functions, initialize r to zero just before the loop
calling cv_timedwait*(), because this code is itself in a loop.

And only return if r == EINTR or r == ERESTART (as returned by
cv_timedwait_sig()), not when r == EWOULDBLOCK.

Those changes seem to fix the crashes in X.Org because the TTM page
fault handler would return VM_PAGER_ERROR, because ttm_bo_wait()
returned an error.