radeon: If PCI expansion ROM is 0, read shadow Video BIOS (at 0xC000)
It appears to be the correct solution even if there's multiple discrete
cards or a combination of integrated + discrete cards.

Even if it's not, it'll allow several users to use the driver for now.

While here, add more log in radeon_bios.c to help debugging in the
future.

Discussed with:	J.R. Oldroyd, avg@